# Use Vector API in the Java Extension