--- a/java/src/json/ext/EscapeScanner.java
+++ b/java/src/json/ext/EscapeScanner.java
@@ -15,33 +15,42 @@ interface EscapeScanner {
     }
 
     static class VectorSupport {
-        static Constructor<?> vectorizedEscapeScannerConstructor = null;
+        static final EscapeScanner VECTORIZED_ESCAPE_SCANNER;
 
         static {
             Optional<Module> vectorModule = ModuleLayer.boot().findModule("jdk.incubator.vector");
+            EscapeScanner scanner = null;
             if (vectorModule.isPresent()) {
                 try {
                     Class<?> vectorEscapeScannerClass = EscapeScanner.class.getClassLoader().loadClass("json.ext.VectorizedEscapeScanner");
-                    vectorizedEscapeScannerConstructor = vectorEscapeScannerClass.getDeclaredConstructor();
-                } catch (ClassNotFoundException | NoSuchMethodException e) {
+                    Constructor<?> vectorizedEscapeScannerConstructor = vectorEscapeScannerClass.getDeclaredConstructor();
+                    scanner = (EscapeScanner) vectorizedEscapeScannerConstructor.newInstance();
+                } catch (ClassNotFoundException | NoSuchMethodException | InstantiationException | IllegalAccessException | InvocationTargetException e) {
                     // Fallback to the ScalarEscapeScanner if we cannot load the VectorizedEscapeScanner.
                     System.err.println("Failed to load VectorizedEscapeScanner, falling back to ScalarEscapeScanner: " + e.getMessage());
+                    scanner = null;
                 }
+                
             }
+            VECTORIZED_ESCAPE_SCANNER = scanner;
         }
     }
 
     boolean scan(EscapeScanner.State state) throws java.io.IOException;
 
-    public static EscapeScanner basicScanner() {
-        if (VectorSupport.vectorizedEscapeScannerConstructor != null) {
-            try {
-                // Attempt to instantiate the vectorized escape scanner if available.
-                return (EscapeScanner) VectorSupport.vectorizedEscapeScannerConstructor.newInstance();
-            } catch (InstantiationException | IllegalAccessException | InvocationTargetException e) {
-                System.err.println("Failed to instantiate VectorizedEscapeScanner, falling back to ScalarEscapeScanner: " + e.getMessage());
-            }
+    default State createState(byte[] ptrBytes, int ptr, int len, int beg) {
+        State state = new State();
+        state.ptrBytes = ptrBytes;
+        state.ptr = ptr;
+        state.len = len;
+        state.beg = beg;
+        state.pos = 0; // Start scanning from the beginning of the segment
+        return state;
+    }
 
+    public static EscapeScanner basicScanner() {
+        if (VectorSupport.VECTORIZED_ESCAPE_SCANNER != null) {
+            return VectorSupport.VECTORIZED_ESCAPE_SCANNER;
         }
 
         return new ScalarEscapeScanner(StringEncoder.ESCAPE_TABLE);