@samyron 2025-07-28
I decided to try a different approach after looking at the HotSpot C2 output. Unlike in the C extension, where we mostly control method inlining, HotSpot isn't so easily influenced. 

I merged `VectorizedStringEncoder` which wraps the escape logic in the vectorized scanning. This reduces method calls back to the search code. 

## Performance of `VectorizedStringEncoder`

```
scott@Scotts-MacBook-Air json % ONLY=json JAVA_OPTS='--add-modules jdk.incubator.vector -Djson.enableVectorizedEscapeScanner=false -Djson.enableVectorizedStringEncoder=true' ruby -I"lib" benchmark/encoder-realworld.rb
WARNING: Using incubator modules: jdk.incubator.vector
VectorizedEscapeScanner disabled.
json.ext.VectorizedStringEncoder loaded successfully.
== Encoding activitypub.json (52595 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a Java HotSpot(TM) 64-Bit Server VM 21.0.7+8-LTS-245 on 21.0.7+8-LTS-245 +jit [arm64-darwin]
Warming up --------------------------------------
                json     1.537k i/100ms
Calculating -------------------------------------
                json     15.382k (± 0.6%) i/s   (65.01 μs/i) -    155.237k in  10.092376s

== Encoding citm_catalog.json (500298 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a Java HotSpot(TM) 64-Bit Server VM 21.0.7+8-LTS-245 on 21.0.7+8-LTS-245 +jit [arm64-darwin]
Warming up --------------------------------------
                json    81.000 i/100ms
Calculating -------------------------------------
                json    818.347 (± 0.7%) i/s    (1.22 ms/i) -      8.181k in   9.997474s

== Encoding twitter.json (466906 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a Java HotSpot(TM) 64-Bit Server VM 21.0.7+8-LTS-245 on 21.0.7+8-LTS-245 +jit [arm64-darwin]
Warming up --------------------------------------
                json   176.000 i/100ms
Calculating -------------------------------------
                json      1.766k (± 1.9%) i/s  (566.28 μs/i) -     17.776k in  10.070684s

== Encoding ohai.json (20147 bytes)
jruby 9.4.12.0 (3.1.4) 2025-02-11 f4ab75096a Java HotSpot(TM) 64-Bit Server VM 21.0.7+8-LTS-245 on 21.0.7+8-LTS-245 +jit [arm64-darwin]
Warming up --------------------------------------
                json     2.426k i/100ms
Calculating -------------------------------------
                json     23.958k (± 0.6%) i/s   (41.74 μs/i) -    240.174k in  10.025043s
```

Additionally, here is a screenshot of VisualVM showing the result of running the `activitypub.json` benchmark for 30 seconds.

<img width="1138" height="636" alt="image" src="https://github.com/user-attachments/assets/a1d80022-04c9-4db0-a393-c5e6374dd488" />