@headius 2025-07-28
@samyron This is interesting progress! I am looking forward to trying it myself now that I'm back in the office.

Yes, HotSpot can be a tricky beast to manipulate. We will want to look at some deeper logging of the JIT and inlining decisions to see whether everything that should be is getting inlined. There's potentially other parts of json unrelated to your changes that are also interfering with inlining (such as the double-dispatching logic to find an appropriate formatter for output text).

Have you tried running on a newer JDK? There's continuous improvements in this area.