@headius 2025-07-28
It's also possible that we are losing too much performance to excessive allocation. I'll try to do some profiling once I get your code up and running.